requirements.txt
boto3
diffusers
fastapi
numpy
Pillow
pydantic
requests
torch
typing-extensions
uvicorn
openai
accelerate
transformers
sentencepiece
python-dotenv
cachetools
xformers
protobuf
python-multipart
python-dotenv
